refactor: make Message component from Messages component
기존의 Messages컴포넌트를 MessageList컴포넌트로 변경 후, Message 컴포넌트로 일부분 추출
